+++ External/FEXCore/Source/CMakeLists.txt
@@ -124,8 +124,6 @@ set (SRCS
   Interface/IR/Passes/SyscallOptimization.cpp
   Utils/Allocator.cpp
   Utils/Allocator/64BitAllocator.cpp
-  Utils/ELFContainer.cpp
-  Utils/ELFSymbolDatabase.cpp
   Utils/LogManager.cpp
   Utils/Threads.cpp
   )
